+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.1
 ---
 
+ * Moved history tracker of out main service into the main controller
  * Fixed issue with quote escaping
  * Changed the way of storing configuration data. Since now its stored in the database
  * Minor improvements in internals